# url-to-code/lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface CloneInput {
  /** The URL of the website to clone */
  url: string;
}

export type CloneJobStatus = typeof CloneJobStatus[keyof typeof CloneJobStatus];


export const CloneJobStatus = {
  pending: 'pending',
  running: 'running',
  done: 'done',
  error: 'error',
} as const;

export interface CloneJob {
  jobId: string;
  status: CloneJobStatus;
}

export type CloneStatusStatus = typeof CloneStatusStatus[keyof typeof CloneStatusStatus];


export const CloneStatusStatus = {
  pending: 'pending',
  running: 'running',
  done: 'done',
  error: 'error',
} as const;

export interface CloneStatus {
  jobId: string;
  status: CloneStatusStatus;
  /** Progress 0-100 */
  progress: number;
  /** @nullable */
  message?: string | null;
  /** @nullable */
  url?: string | null;
  downloadReady?: boolean;
}

export interface ErrorResponse {
  error: string;
}