# url-to-code/lib/api-zod/src/generated/types/cloneInput.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface CloneInput {
  /** The URL of the website to clone */
  url: string;
}