# url-to-code/lib/api-zod/src/generated/types/cloneJob.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { CloneJobStatus } from './cloneJobStatus';

export interface CloneJob {
  jobId: string;
  status: CloneJobStatus;
}