# url-to-code/lib/api-zod/src/generated/types/cloneStatus.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { CloneStatusStatus } from './cloneStatusStatus';

export interface CloneStatus {
  jobId: string;
  status: CloneStatusStatus;
  /** Progress 0-100 */
  progress: number;
  /** @nullable */
  message?: string | null;
  /** @nullable */
  url?: string | null;
  downloadReady?: boolean;
}